Refactor & change long, lat data type to double

+++ src/Tests/UnitTests/Weather.Infrastructure.UnitTests/Services/WeatherServiceTests.cs
@@ -28,14 +28,14 @@ namespace Weather.Infrastructure.UnitTests.Services
             var location = new LocationDto { Latitude = 15, Longitude = 25 };
             var failedMessage = "message";
 
-            _weatherbiClientMock.Setup(x=>x.GetCurrentWeather(It.IsAny<long>(), It.IsAny<long>(), It.IsAny<CancellationToken>())).ReturnsAsync(Result.Fail(failedMessage));
+            _weatherbiClientMock.Setup(x=>x.GetCurrentWeather(It.IsAny<double>(), It.IsAny<double>(), It.IsAny<CancellationToken>())).ReturnsAsync(Result.Fail(failedMessage));
             //Act
             var result = await _uut.GetCurrentWeather(location, CancellationToken.None);
             //Assert
             Assert.True(result.IsFailed);
             Assert.Single(result.Errors);
             Assert.Equal(failedMessage, result.Errors.First().Message);
-            _weatherbiClientMock.Verify(x => x.GetCurrentWeather(It.Is<long>(y=>y.Equals(location.Latitude)), It.Is<long>(y => y.Equals(location.Longitude)), It.IsAny<CancellationToken>()), Times.Once);
+            _weatherbiClientMock.Verify(x => x.GetCurrentWeather(It.Is<double>(y=>y.Equals(location.Latitude)), It.Is<double>(y => y.Equals(location.Longitude)), It.IsAny<CancellationToken>()), Times.Once);
         }
 
         [Fact]
@@ -44,14 +44,14 @@ namespace Weather.Infrastructure.UnitTests.Services
             //Arrange
             var location = new LocationDto { Latitude = 15, Longitude = 25 };
 
-            _weatherbiClientMock.Setup(x => x.GetCurrentWeather(It.IsAny<long>(), It.IsAny<long>(), It.IsAny<CancellationToken>())).ReturnsAsync(Result.Ok((CurrentWeatherDataDto)null));
+            _weatherbiClientMock.Setup(x => x.GetCurrentWeather(It.IsAny<double>(), It.IsAny<double>(), It.IsAny<CancellationToken>())).ReturnsAsync(Result.Ok((CurrentWeatherDataDto)null));
             //Act
             var result = await _uut.GetCurrentWeather(location, CancellationToken.None);
             //Assert
             Assert.True(result.IsFailed);
             Assert.Single(result.Errors);
             Assert.Equal(ErrorMessages.ExternalClientGetDataFailed_EmptyOrNull, result.Errors.First().Message);
-            _weatherbiClientMock.Verify(x => x.GetCurrentWeather(It.Is<long>(y => y.Equals(location.Latitude)), It.Is<long>(y => y.Equals(location.Longitude)), It.IsAny<CancellationToken>()), Times.Once);
+            _weatherbiClientMock.Verify(x => x.GetCurrentWeather(It.Is<double>(y => y.Equals(location.Latitude)), It.Is<double>(y => y.Equals(location.Longitude)), It.IsAny<CancellationToken>()), Times.Once);
         }
 
         [Fact]
@@ -60,14 +60,14 @@ namespace Weather.Infrastructure.UnitTests.Services
             //Arrange
             var location = new LocationDto { Latitude = 15, Longitude = 25 };
 
-            _weatherbiClientMock.Setup(x => x.GetCurrentWeather(It.IsAny<long>(), It.IsAny<long>(), It.IsAny<CancellationToken>())).ReturnsAsync(Result.Ok(new CurrentWeatherDataDto()));
+            _weatherbiClientMock.Setup(x => x.GetCurrentWeather(It.IsAny<double>(), It.IsAny<double>(), It.IsAny<CancellationToken>())).ReturnsAsync(Result.Ok(new CurrentWeatherDataDto()));
             //Act
             var result = await _uut.GetCurrentWeather(location, CancellationToken.None);
             //Assert
             Assert.True(result.IsFailed);
             Assert.Single(result.Errors);
             Assert.Contains(ErrorMessages.ExternalClientGetDataFailed_EmptyOrNull, result.Errors.First().Message);
-            _weatherbiClientMock.Verify(x => x.GetCurrentWeather(It.Is<long>(y => y.Equals(location.Latitude)), It.Is<long>(y => y.Equals(location.Longitude)), It.IsAny<CancellationToken>()), Times.Once);
+            _weatherbiClientMock.Verify(x => x.GetCurrentWeather(It.Is<double>(y => y.Equals(location.Latitude)), It.Is<double>(y => y.Equals(location.Longitude)), It.IsAny<CancellationToken>()), Times.Once);
         }
 
         [Fact]
@@ -81,7 +81,7 @@ namespace Weather.Infrastructure.UnitTests.Services
                 new Wheaterbit.Client.Dtos.CurrentWeatherDto()
             };
 
-            _weatherbiClientMock.Setup(x => x.GetCurrentWeather(It.IsAny<long>(), It.IsAny<long>(), It.IsAny<CancellationToken>())).ReturnsAsync(Result.Ok(new CurrentWeatherDataDto
+            _weatherbiClientMock.Setup(x => x.GetCurrentWeather(It.IsAny<double>(), It.IsAny<double>(), It.IsAny<CancellationToken>())).ReturnsAsync(Result.Ok(new CurrentWeatherDataDto
             { 
                 Data = data
             }));
@@ -91,7 +91,7 @@ namespace Weather.Infrastructure.UnitTests.Services
             Assert.True(result.IsFailed);
             Assert.Single(result.Errors);
             Assert.Contains(string.Format(ErrorMessages.ExternalClientGetDataFailed_CorruptedData_InvalidCount, 2), result.Errors.First().Message);
-            _weatherbiClientMock.Verify(x => x.GetCurrentWeather(It.Is<long>(y => y.Equals(location.Latitude)), It.Is<long>(y => y.Equals(location.Longitude)), It.IsAny<CancellationToken>()), Times.Once);
+            _weatherbiClientMock.Verify(x => x.GetCurrentWeather(It.Is<double>(y => y.Equals(location.Latitude)), It.Is<double>(y => y.Equals(location.Longitude)), It.IsAny<CancellationToken>()), Times.Once);
         }
 
         [Fact]
@@ -106,7 +106,7 @@ namespace Weather.Infrastructure.UnitTests.Services
 
             var mapResult = new Domain.Dtos.CurrentWeatherDto();
 
-            _weatherbiClientMock.Setup(x => x.GetCurrentWeather(It.IsAny<long>(), It.IsAny<long>(), It.IsAny<CancellationToken>())).ReturnsAsync(Result.Ok(new CurrentWeatherDataDto
+            _weatherbiClientMock.Setup(x => x.GetCurrentWeather(It.IsAny<double>(), It.IsAny<double>(), It.IsAny<CancellationToken>())).ReturnsAsync(Result.Ok(new CurrentWeatherDataDto
             {
                 Data = data
             }));
@@ -119,7 +119,7 @@ namespace Weather.Infrastructure.UnitTests.Services
             //Assert
             Assert.True(result.IsSuccess);
             Assert.Empty(result.Errors);
-            _weatherbiClientMock.Verify(x => x.GetCurrentWeather(It.Is<long>(y => y.Equals(location.Latitude)), It.Is<long>(y => y.Equals(location.Longitude)), It.IsAny<CancellationToken>()), Times.Once);
+            _weatherbiClientMock.Verify(x => x.GetCurrentWeather(It.Is<double>(y => y.Equals(location.Latitude)), It.Is<double>(y => y.Equals(location.Longitude)), It.IsAny<CancellationToken>()), Times.Once);
             Assert.Equivalent(mapResult, result.Value);
             _mapperMock.Verify(x => x.Map<Domain.Dtos.CurrentWeatherDto>(It.IsAny<Wheaterbit.Client.Dtos.CurrentWeatherDto>()), Times.Once);
         }
@@ -131,14 +131,14 @@ namespace Weather.Infrastructure.UnitTests.Services
             var location = new LocationDto { Latitude = 15, Longitude = 25 };
             var failedMessage = "message";
 
-            _weatherbiClientMock.Setup(x => x.GetSixteenDayForecast(It.IsAny<long>(), It.IsAny<long>(), It.IsAny<CancellationToken>())).ReturnsAsync(Result.Fail(failedMessage));
+            _weatherbiClientMock.Setup(x => x.GetSixteenDayForecast(It.IsAny<double>(), It.IsAny<double>(), It.IsAny<CancellationToken>())).ReturnsAsync(Result.Fail(failedMessage));
             //Act
             var result = await _uut.GetForecastWeather(location, CancellationToken.None);
             //Assert
             Assert.True(result.IsFailed);
             Assert.Single(result.Errors);
             Assert.Equal(failedMessage, result.Errors.First().Message);
-            _weatherbiClientMock.Verify(x => x.GetSixteenDayForecast(It.Is<long>(y => y.Equals(location.Latitude)), It.Is<long>(y => y.Equals(location.Longitude)), It.IsAny<CancellationToken>()), Times.Once);
+            _weatherbiClientMock.Verify(x => x.GetSixteenDayForecast(It.Is<double>(y => y.Equals(location.Latitude)), It.Is<double>(y => y.Equals(location.Longitude)), It.IsAny<CancellationToken>()), Times.Once);
         }
 
         [Fact]
@@ -147,14 +147,14 @@ namespace Weather.Infrastructure.UnitTests.Services
             //Arrange
             var location = new LocationDto { Latitude = 15, Longitude = 25 };
 
-            _weatherbiClientMock.Setup(x => x.GetSixteenDayForecast(It.IsAny<long>(), It.IsAny<long>(), It.IsAny<CancellationToken>())).ReturnsAsync(Result.Ok((Wheaterbit.Client.Dtos.ForecastWeatherDto)null));
+            _weatherbiClientMock.Setup(x => x.GetSixteenDayForecast(It.IsAny<double>(), It.IsAny<double>(), It.IsAny<CancellationToken>())).ReturnsAsync(Result.Ok((Wheaterbit.Client.Dtos.ForecastWeatherDto)null));
             //Act
             var result = await _uut.GetForecastWeather(location, CancellationToken.None);
             //Assert
             Assert.True(result.IsFailed);
             Assert.Single(result.Errors);
             Assert.Equal(ErrorMessages.ExternalClientGetDataFailed_EmptyOrNull, result.Errors.First().Message);
-            _weatherbiClientMock.Verify(x => x.GetSixteenDayForecast(It.Is<long>(y => y.Equals(location.Latitude)), It.Is<long>(y => y.Equals(location.Longitude)), It.IsAny<CancellationToken>()), Times.Once);
+            _weatherbiClientMock.Verify(x => x.GetSixteenDayForecast(It.Is<double>(y => y.Equals(location.Latitude)), It.Is<double>(y => y.Equals(location.Longitude)), It.IsAny<CancellationToken>()), Times.Once);
         }
 
         [Fact]
@@ -163,14 +163,14 @@ namespace Weather.Infrastructure.UnitTests.Services
             //Arrange
             var location = new LocationDto { Latitude = 15, Longitude = 25 };
 
-            _weatherbiClientMock.Setup(x => x.GetSixteenDayForecast(It.IsAny<long>(), It.IsAny<long>(), It.IsAny<CancellationToken>())).ReturnsAsync(Result.Ok(new Wheaterbit.Client.Dtos.ForecastWeatherDto()));
+            _weatherbiClientMock.Setup(x => x.GetSixteenDayForecast(It.IsAny<double>(), It.IsAny<double>(), It.IsAny<CancellationToken>())).ReturnsAsync(Result.Ok(new Wheaterbit.Client.Dtos.ForecastWeatherDto()));
             //Act
             var result = await _uut.GetForecastWeather(location, CancellationToken.None);
             //Assert
             Assert.True(result.IsFailed);
             Assert.Single(result.Errors);
             Assert.Contains(ErrorMessages.ExternalClientGetDataFailed_EmptyOrNull, result.Errors.First().Message);
-            _weatherbiClientMock.Verify(x => x.GetSixteenDayForecast(It.Is<long>(y => y.Equals(location.Latitude)), It.Is<long>(y => y.Equals(location.Longitude)), It.IsAny<CancellationToken>()), Times.Once);
+            _weatherbiClientMock.Verify(x => x.GetSixteenDayForecast(It.Is<double>(y => y.Equals(location.Latitude)), It.Is<double>(y => y.Equals(location.Longitude)), It.IsAny<CancellationToken>()), Times.Once);
         }
 
         [Fact]
@@ -188,7 +188,7 @@ namespace Weather.Infrastructure.UnitTests.Services
 
             var mapResult = new Domain.Dtos.ForecastWeatherDto();
 
-            _weatherbiClientMock.Setup(x => x.GetSixteenDayForecast(It.IsAny<long>(), It.IsAny<long>(), It.IsAny<CancellationToken>())).ReturnsAsync(Result.Ok(data));
+            _weatherbiClientMock.Setup(x => x.GetSixteenDayForecast(It.IsAny<double>(), It.IsAny<double>(), It.IsAny<CancellationToken>())).ReturnsAsync(Result.Ok(data));
 
             _mapperMock.Setup(x => x.Map<Domain.Dtos.ForecastWeatherDto>(It.IsAny<Wheaterbit.Client.Dtos.ForecastWeatherDto>())).Returns(mapResult);
 
@@ -198,7 +198,7 @@ namespace Weather.Infrastructure.UnitTests.Services
             //Assert
             Assert.True(result.IsSuccess);
             Assert.Empty(result.Errors);
-            _weatherbiClientMock.Verify(x => x.GetSixteenDayForecast(It.Is<long>(y => y.Equals(location.Latitude)), It.Is<long>(y => y.Equals(location.Longitude)), It.IsAny<CancellationToken>()), Times.Once);
+            _weatherbiClientMock.Verify(x => x.GetSixteenDayForecast(It.Is<double>(y => y.Equals(location.Latitude)), It.Is<double>(y => y.Equals(location.Longitude)), It.IsAny<CancellationToken>()), Times.Once);
             Assert.Equivalent(mapResult, result.Value);
             _mapperMock.Verify(x => x.Map<Domain.Dtos.ForecastWeatherDto>(It.IsAny<Wheaterbit.Client.Dtos.ForecastWeatherDto>()), Times.Once);
         }
