+++ src/Weather.Domain/Queries/GetCurrentWeatherQuery.cs
@@ -5,7 +5,7 @@ namespace Weather.Domain.Queries
     public sealed class GetCurrentWeatherQuery
     {
         public LocationDto Location { get; init; }
-        public GetCurrentWeatherQuery(long latitude, long longtitude)
+        public GetCurrentWeatherQuery(double latitude, double longtitude)
         {
             Location = new LocationDto
             {
