+++ src/Wheaterbit.Client/Abstractions/IWeatherbitHttpClient.cs
@@ -5,7 +5,7 @@ namespace Wheaterbit.Client.Abstractions
 {
     public interface IWeatherbitHttpClient
     {
-        Task<Result<ForecastWeatherDto>> GetSixteenDayForecast(long latitude, long longitude, CancellationToken cancellationToken);
-        Task<Result<CurrentWeatherDataDto>> GetCurrentWeather(long latitude, long longitude, CancellationToken cancellationToken);
+        Task<Result<ForecastWeatherDto>> GetSixteenDayForecast(double latitude, double longitude, CancellationToken cancellationToken);
+        Task<Result<CurrentWeatherDataDto>> GetCurrentWeather(double latitude, double longitude, CancellationToken cancellationToken);
     }
 }
